+++ src/sbml-export/index.js
@@ -30,8 +30,6 @@ class SBMLExport extends AbstractExport {
       this.spaceFilter = q.spaceFilter;
     } else if (typeof q.spaceFilter === 'string') {
       this.spaceFilter = [q.spaceFilter];
-    } else {
-      this.spaceFilter = ['nameless'];
     }
   }
   get className(){
@@ -44,31 +42,50 @@ class SBMLExport extends AbstractExport {
     return ajv.compile(schema);
   }
   make(){
-    // use only one namespace
     let logger = this._container.logger;
-    if (this.spaceFilter.length === 0) {
-      let msg = 'spaceFilter for SBML format should include at least one namespace but get empty';
-      logger.error(msg);
-      var content = '';
-    } else if (!this._container.namespaceStorage.has(this.spaceFilter[0])) {
-      let msg = `Namespace "${this.spaceFilter[0]}" does not exist.`;
-      logger.error(msg);
-      content = '';
-    } else {
-      if (this.spaceFilter.length > 1) {
-        let msg = `SBML format does not support multispace export. Only first namespace "${this.spaceFilter[0]}" will be used.`;
-        logger.warn(msg);
+
+    if (this.spaceFilter !== undefined) {
+      // empty namespace is not allowed
+      if (this.spaceFilter.length === 0) {
+        let msg = 'spaceFilter for SBML format should include at least one namespace, got empty';
+        logger.error(msg);
+        return []; // BRAKE
       }
-      let ns = this._container.namespaceStorage.get(this.spaceFilter[0]);
-      let image = this.getSBMLImage(ns);
-      content = this.getSBMLCode(image);
+      
+      // check if namespaces exists
+      let lostNamespaces = this.spaceFilter.filter((x) => !this._container.namespaceStorage.has(x));
+      if (lostNamespaces.length > 0) {
+        let msg = `Namespaces: ${lostNamespaces.join(', ')} do not exist. SBML export stopped.`;
+        logger.error(msg);
+        return []; // BRAKE
+      }
     }
 
-    return [{
-      content: content,
-      pathSuffix: '/model.xml',
-      type: 'text'
-    }];
+    // filter namespaces if set
+    let selectedNamespaces = this.spaceFilter !== undefined 
+      ? [...this._container.namespaceStorage].filter((x) => this.spaceFilter.indexOf(x[0]) !== -1)
+      : [...this._container.namespaceStorage];
+    
+    let results = selectedNamespaces.map((x) => {
+      let spaceName = x[0];
+      let ns = x[1];
+      let image = this.getSBMLImage(ns);
+      var content = this.getSBMLCode(image);
+
+      if (ns.isAbstract) {
+        let msg = `UnitDefinitions in SBML will be skipped for the abstract namespace "${spaceName}".`;
+        logger.warn(msg);
+      }
+      
+      return {
+        content: content,
+        pathSuffix: `/${spaceName}.xml`,
+        type: 'text'
+      };
+    });
+
+
+    return results;
   }
   getSBMLImage(ns){
     let logger = ns.container.logger;